+++ docs/CHANGELOG.md
@@ -15,6 +15,8 @@
   * `InspectVariableInScope` and `InspectVariable` renamed to `GetVariable` and `GetVariableByScopeAndName`
     respectively. Alias to the older names left for compatibility (and marked as deprecated), but they
     will be removed in future versions.
+* Package `tensors`:
+  * Added `Tensor.Summary(precision int)` to pretty-print a summary of the values of a tensor, numpy-like.
 
 ## v0.14.0 - 2024/10/24